Refactor type usage across multiple files for improved 64-bit compatibility
- Updated various loops and comparisons to use `size_t` instead of `int`, enhancing safety and compatibility with larger datasets.
- Adjusted string length handling to utilize `ssize_t` where appropriate, preventing potential data loss.
- Documented type usage guidelines in AGENTS.md, clarifying the distinctions between `Int`, `ssize_t`, and `CLogicInt` for new code.
- Ensured consistent type handling across the codebase to support 64-bit architecture.

+++ sakura_core/view/CEditView_Paint.cpp
@@ -1083,7 +1083,7 @@ bool CEditView::DrawLayoutLine(SColorStrategyInfo* pInfo)
 			nPosLength = nPosInLogic - nPosBgn;
 			//1文字情報取得
 			CFigure& cFigure = pcFigureManager->GetFigure(&cLineStr.GetPtr()[nPosInLogic],
-				static_cast<int>(cLineStr.GetLength() - nPosInLogic));
+				cLineStr.GetLength() - nPosInLogic);
 			FigureRenderType nextRenderType = CFigure_Text::RenderType_None;
 			bool is_text = (typeid(cFigure) == typeid(CFigure_Text));
 			if (is_text) {
