+++ sakura_core/view/figures/CFigure_CtrlCode.cpp
@@ -14,7 +14,7 @@
 //                     CFigure_CtrlCode                        //
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
 
-bool CFigure_CtrlCode::Match(const wchar_t* pText, [[maybe_unused]] int nTextLen) const
+bool CFigure_CtrlCode::Match(const wchar_t* pText, [[maybe_unused]] ssize_t nTextLen) const
 {
 	//当面はASCII制御文字（C0 Controls, IsHankaku()で半角扱い）だけを制御文字表示にする
 	//そうしないと IsHankaku(0x0600)==false なのに iswcntrl(0x0600)!=0 のようなケースで表示桁がずれる
