+++ sakura_core/view/figures/CFigure_HanSpace.cpp
@@ -15,7 +15,7 @@
 //                     CFigure_HanSpace                        //
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
 
-bool CFigure_HanSpace::Match(const wchar_t* pText, [[maybe_unused]] int nTextLen) const
+bool CFigure_HanSpace::Match(const wchar_t* pText, [[maybe_unused]] ssize_t nTextLen) const
 {
 	if( pText[0] == L' ' ){
 		return true;
