+++ sakura_core/view/figures/CFigure_Tab.h
@@ -15,7 +15,7 @@
 class CFigure_Tab final : public CFigureSpace{
 public:
 	//traits
-	bool Match(const wchar_t* pText, int nTextLen) const override;
+	bool Match(const wchar_t* pText, ssize_t nTextLen) const override;
 	bool Disp(void) const override
 	{
 		return true;
